requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# LangChain dependencies
langchain==0.0.352
langchain-community==0.0.38

# Web interface dependencies
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1

# Security dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Monitoring and logging
structlog==23.2.0
rich==13.7.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Optional: Redis for caching (uncomment if needed)
# redis==5.0.1
# aioredis==2.0.1